requirements.txt
pandas
mysql-connector-python
flask
werkzeug